Feat/ajustes finais

+++ src/main/java/com/example/cmseventosapi/Services/EditionService.java
@@ -1,12 +1,19 @@
 package com.example.cmseventosapi.Services;
 
+import com.example.cmseventosapi.Auth.Exceptions.UserMustBeAdminToPerformActionException;
+import com.example.cmseventosapi.Model.Event;
+import com.example.cmseventosapi.Model.Requests.CreateEditionReq;
+import com.example.cmseventosapi.Repositories.EventRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import com.example.cmseventosapi.Model.Edition;
 import com.example.cmseventosapi.Model.User;
 import com.example.cmseventosapi.Repositories.EditionRepository;
 import com.example.cmseventosapi.Repositories.UserRepository;
+import org.webjars.NotFoundException;
 
 @Service
 public class EditionService {
@@ -14,11 +21,29 @@ public class EditionService {
     @Autowired
     private EditionRepository repository;
 
-    public Edition CreateEdition(Edition edition) {
+    @Autowired
+    private EventRepository eventRepository;
+
+    public Edition CreateEdition(CreateEditionReq editionReq) {
+        Event event = this.eventRepository.findById(editionReq.getEventId())
+                .orElseThrow(() -> new NotFoundException("Event not found"));
+
+        Edition edition = new Edition();
+        edition.setNumber(editionReq.getNumber());
+        edition.setYear(editionReq.getYear());
+        edition.setStartDate(editionReq.getStartDate());
+        edition.setEndDate(editionReq.getEndDate());
+        edition.setEvent(event);
+
         return this.repository.save(edition);
     }
 
     public Edition UpdateEdition(Edition edition, Long id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User loggedUser = this.userRepository.findByLogin(authentication.getName()).get();
+        if(!loggedUser.isAdmin()){
+            throw new UserMustBeAdminToPerformActionException("You don't have permission to do this");
+        }
         Edition editionToUpdate = this.repository.findById(id).get();
         editionToUpdate.setNumber(edition.getNumber());
         editionToUpdate.setYear(edition.getYear());
@@ -29,17 +54,29 @@ public class EditionService {
         return this.repository.save(edition);
     }
 
+    
+
     public Edition GetEdition(Long id) {
         return this.repository.findById(id).get();
     }
 
     public void DeleteEdition(Long id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User loggedUser = this.userRepository.findByLogin(authentication.getName()).get();
+        if(!loggedUser.isAdmin()){
+            throw new UserMustBeAdminToPerformActionException("You don't have permission to do this");
+        }
         this.repository.deleteById(id);
     }
 
     @Autowired
     private UserRepository userRepository;
-    public void addOrganizerToEdition(Long editionId, Long userId) {
+    public Edition addOrganizerToEdition(Long editionId, Long userId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User loggedUser = this.userRepository.findByLogin(authentication.getName()).get();
+        if(!loggedUser.isAdmin()){
+            throw new UserMustBeAdminToPerformActionException("You don't have permission to do this");
+        }
         Edition edition = this.repository.findById(editionId).orElseThrow(() -> new RuntimeException("Edição não encontrada"));
         User user = this.userRepository.findById(userId).orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
 
@@ -47,6 +84,6 @@ public class EditionService {
         edition.setOrganizer(user);
 
         this.userRepository.save(user);
-        this.repository.save(edition);
+        return this.repository.save(edition);
     }
 }
